Don't translate :multi-valued facets in Facet#value.
In order to process facet value/counts for :multi-valued facets, one must not attempt to translate them, otherwise we'll end up with one of the object.send()s in #translate returning an Array of values.  A much simpler solution is to just use the attribute_value directly.  Unfortunately, we're then left only being able to use integral :multi values in our facets.