Add total_entries attribute to FacetCollection hash.
When Foo.facets results are collated, build a Hash of total_entries values as returned by Sphinx and then make it available on the FacetCollection. If we then call f = Asset.facets :limit => 100 then f.total_entries is a Hash from facet names to the total number of facet groups available on that facet.